requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
palworld-api==2.0.0
gamercon-async==1.0.7
aiohttp==3.11.18
aiosqlite==0.21.0
aiocache==0.12.3
paramiko==3.4.0